+++ apps/web/src/components/layouts/setting-menu/SettingMenuBox.tsx
@@ -39,10 +39,12 @@ const SettingMenuBox: React.FC<IBoxProps> = ({
       show={show}
       className={`${boxCN} ${flexCN}`}
       enter="transition-all duration-500"
-      enterTo={'px-3 opacity-100 overflow-hidden ' + transitionFlexCN}
+      enterTo={['px-3 opacity-100 overflow-hidden', transitionFlexCN].join(' ')}
       enterFrom="flex-[0] px-0 opacity-0 overflow-hidden"
       leave="transition-all duration-500"
-      leaveFrom={'px-3 opacity-100 overflow-hidden ' + transitionFlexCN}
+      leaveFrom={['px-3 opacity-100 overflow-hidden', transitionFlexCN].join(
+        ' ',
+      )}
       leaveTo="flex-[0] px-0 opacity-0 overflow-hidden"
     >
       {children}
