+++ apps/web/src/containers/setting-menu/FieldSetting/OptionBadge.tsx
@@ -66,7 +66,7 @@ const OptionDeletePopover: React.FC<IProps> = ({
           {t('main.setting.dialog.delete-option.title')}
         </PopoverHeading>
         <div className="m-5">
-          <p className={'font-14-regular mb-10 whitespace-pre-line'}>
+          <p className="font-14-regular mb-10 whitespace-pre-line">
             {t('main.setting.dialog.delete-option.description')}
           </p>
           <div className="flex justify-end gap-2">
@@ -74,8 +74,8 @@ const OptionDeletePopover: React.FC<IProps> = ({
               {t('button.cancel')}
             </button>
             <button
-              className={'btn btn-primary'}
-              type={'button'}
+              className="btn btn-primary"
+              type="button"
               onClick={() => {
                 if (typeof targetOptionIndex !== 'undefined') {
                   removeOption(targetOptionIndex);
